+++ source/rr/src/gameexec.cpp
@@ -1100,7 +1100,7 @@ static int32_t VM_ResetPlayer(int const playerNum, int32_t vmFlags)
         if (playerNum == myconnectindex)
         {
             CAMERADIST = 0;
-            CAMERACLOCK = totalclock;
+            CAMERACLOCK = (int32_t)totalclock;
         }
 
         if (g_fakeMultiMode)
