+++ source/rr/src/savegame.cpp
@@ -419,9 +419,9 @@ static struct {
 
 static void G_SaveTimers(void)
 {
-    g_timers.totalclock     = totalclock;
-    g_timers.totalclocklock = totalclocklock;
-    g_timers.ototalclock    = ototalclock;
+    g_timers.totalclock     = (int32_t)totalclock;
+    g_timers.totalclocklock = (int32_t)totalclocklock;
+    g_timers.ototalclock    = (int32_t)ototalclock;
     g_timers.lockclock      = lockclock;
 }
